+++ kern/src/ioapic.rs
@@ -10,19 +10,19 @@ use x86::*;
 
 pub const IOAPIC: usize = 0xFEC00000; // Default physical address of IO APIC
 
-pub const REG_ID: u32 = 0x00; // Register index: ID
-pub const REG_VER: u32 = 0x01; // Register index: version
-pub const REG_TABLE: u32 = 0x10; // Redirection table base
+pub const REG_ID: usize = 0x00; // Register index: ID
+pub const REG_VER: usize = 0x01; // Register index: version
+pub const REG_TABLE: usize = 0x10; // Redirection table base
 
 // The redirection table starts at REG_TABLE and uses
 // two registers to configure each interrupt.
 // The first (low) register in a pair contains configuration bits.
 // The second (high) register contains a bitmask telling which
 // CPUs can serve that interrupt.
-pub const INT_DISABLED: u32 = 0x00010000; // Interrupt disabled
-pub const INT_LEVEL: u32 = 0x00008000; // Level-triggered (vs edge-)
-pub const INT_ACTIVELOW: u32 = 0x00002000; // Active low (vs high)
-pub const INT_LOGICAL: u32 = 0x00000800; // Destination is CPU id (vs APIC ID)
+pub const INT_DISABLED: usize = 0x00010000; // Interrupt disabled
+pub const INT_LEVEL: usize = 0x00008000; // Level-triggered (vs edge-)
+pub const INT_ACTIVELOW: usize = 0x00002000; // Active low (vs high)
+pub const INT_LOGICAL: usize = 0x00000800; // Destination is CPU id (vs APIC ID)
 
 // TODO: volatile
 static mut ioapic: *mut Ioapic = core::ptr::null_mut();
@@ -30,17 +30,17 @@ static mut ioapic: *mut Ioapic = core::ptr::null_mut();
 // IO APIC MMIO structure: write reg, then read or write data.
 #[repr(C)]
 struct Ioapic {
-    reg: u32,
-    pad: [u32; 3],
-    data: u32,
+    reg: usize,
+    pad: [usize; 3],
+    data: usize,
 }
 
-pub unsafe fn ioapicread(reg: u32) -> u32 {
+pub unsafe fn ioapicread(reg: usize) -> usize {
     (*ioapic).reg = reg;
     (*ioapic).data
 }
 
-pub unsafe fn ioapicwrite(reg: u32, data: u32) {
+pub unsafe fn ioapicwrite(reg: usize, data: usize) {
     (*ioapic).reg = reg;
     (*ioapic).data = data;
 }
@@ -65,7 +65,7 @@ pub unsafe fn ioapicinit() {
     }
 }
 
-pub unsafe fn ioapicenable(irq: u32, cpunum: u32) {
+pub unsafe fn ioapicenable(irq: usize, cpunum: usize) {
     if (!ismp) {
         return;
     }
